+++ manage.py
@@ -3,9 +3,9 @@ import os
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
 
-from app import db, create_app
+from app import db
+from app.views import app
 
-app = create_app(config_name=os.getenv('APP_SETTINGS'))
 migrate = Migrate(app, db)
 manager = Manager(app)
